+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Development
 * Add more tests
 * Fix fetch_all() – use "next" value #47
+* Fix to_resource, tests and readme – replace "id" search param by "_id" #55
 
 ## 1.1.0
 * Remove schemas and resource key validation #33