+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add more tests
 * Fix fetch_all() – use "next" value #47
 * Fix to_resource, tests and readme – replace "id" search param by "_id" #55
+* Fix for absolute url's in "next" link (fetch_all/searchset iterator). Related to #47
 
 ## 1.1.0
 * Remove schemas and resource key validation #33